requirements.txt
flask>=3.0
gunicorn>=21.0
python-dotenv>=1.0
langchain>=0.3
langchain-openai>=0.2
langchain-community>=0.3
langgraph>=0.2
duckduckgo-search>=5.0,<6.0
numexpr>=2.9
